Fix repeated calling of contact link removal
this._check_edge_for_removal(this_edg) in simulation.js is pushed to
event_queue twice for each link, because each link has two nodes, and the
function will be pushed by both nodes in one collision event. I changed
the push operation to be under if condition, so that the pushing will
only occur once.